internal/reflectlite, reflect: move StringHeader and SliceHeader back to package reflect
